requirements.txt
python-telegram-bot>=21.0,<22.0
aiosqlite>=0.20.0,<1.0
APScheduler>=3.10.4,<4.0
python-dotenv>=1.0.1,<2.0

